requirements.txt
# Core ML Libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# AutoGluon Ensemble
autogluon.tabular>=0.8.2

# Data Processing
PyYAML>=6.0

# Logging and utilities
tqdm>=4.64.0

# Optional: For visualization and analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# For high-performance computing (optional)
numba>=0.56.0 